tests/test_app.py
import falcon
from falcon import testing
import msgpack
import pytest

from eddie.app import api


@pytest.fixture
def client():
    return testing.TestClient(api)



def test_list_images(client):
    doc = {
        'images': [
            {
                'href': '/images/1eaf6ef1-7f2d-4ecc-a8d5-6e8adba7cc0e.png'
            }
        ]
    }

    response = client.simulate_get('/quotes')
    result_doc = msgpack.unpackb(response.content, encoding='utf-8')

    assert result_doc == doc
    assert response.status == falcon.HTTP_OK