requirements.txt
appdirs==1.4.3
certifi==2017.11.5
cffi==1.11.1
chardet==3.0.4
cycler==0.10.0
falcon==1.3.0
geopy==1.11.0
greenlet==0.4.12
gunicorn==19.7.1
idna==2.6
jsonschema==2.6.0
msgpack-python==0.4.8
numpy==1.13.3
packaging==16.8
py==1.5.2
pymongo==3.5.1
pyparsing==2.2.0
pytest==3.2.5
python-dateutil==2.6.1
python-mimeparse==1.6.0
pytz==2017.3
readline==6.2.4.1
requests==2.18.4
rethinkdb==2.3.0.post6
six==1.11.0
urllib3==1.22
